refactor(B-INT-7): decompose bicameral-card.js + MarketplaceRoute.ts under razor
bicameral-card.js 314 -> 98 (state renderers -> bicameral-card-render.js);
MarketplaceRoute.ts 382 -> 29 (shared HITL-nonce core + read/install/scan modules).
Follow-up (#413): marketplace install/:id/confirm completion decomposed into
handleInstallCompletion/runPostInstallScan/recordInstallLedger to clear the
40-line function razor. Public surfaces preserved; zero behavioral change:
90 mocha + 10 Playwright verbatim. META_LEDGER #411 + #413.
